haywire will now keep allocating memory as required, as it's not guaranteed that we'll get one and only one fully formed request on each combination of alloc/read calls.
Previously, we were assuming that the alloc callback was only called once per request and that the read callback was called when all the data was available.
Unfortunately, that's not the case, and the read callback may be called when only a fraction of the incoming data is available, especially under high concurrency.

Buffers are now managed in http_request_buffers.*. We use a mark/sweep technique to get rid of requests previously handled by a connection whenever a new buffer chunk is fully processed. This allows us to keep a relatively small buffer. Also, it's now possible to register interest in a given memory region by placing a pin on it and retrieving its location once the request has been fully read in.

We will now also deal with bad requests that were causing crashes before, requests that are too long to be processed and that could cause us to run out of memory. Unknown errors are also dealt with now, by returning an appropriate error code.

Removed redundant header imports.